added Config.jsx, created functions in Map and List Screens to fetch data from realtime database
