+++ foodtruckapp/Components/MapScreen.js
@@ -1,10 +1,10 @@
 import { StyleSheet, View } from 'react-native';
 import * as React from 'react';
 import MapView, { PROVIDER_GOOGLE, Marker } from 'react-native-maps';
-import Geocoder from 'react-native-geocoding';
 import Nav from './Nav';
 import Header from './Header';
-import fetchTrucks from './Fetch';
+import { db } from './Config';
+import { ref, get, child } from 'firebase/database';
 
 export default function MapScreen({ navigation }) {  
 
@@ -173,11 +173,31 @@ export default function MapScreen({ navigation }) {
       ];
 
     const [truckData, setTruckData] = React.useState([]);
+    const [loading, setLoading] = React.useState(false);
+  
     React.useEffect(() => {
       const unsubscribe = navigation.addListener('focus', () => {
         // The screen is focused
         // Call any action
-        setTruckData(fetchTrucks());
+        setTimeout(() => {
+          const dbRef = ref(db);
+          get(child(dbRef, `users/82LyYqZ73TZ2XUZizHj9piktknm1/data`)).then((snapshot) => {
+              if (snapshot.exists()) {
+                  const data = snapshot.val();
+                  const trucks = Object.keys(data).map(key => ({
+                      ...data[key]
+                  }))
+                  const sortedTruckData = trucks.sort((a, b) => a.name.localeCompare(b.name));
+                  setTruckData(sortedTruckData);
+              } else {
+                  console.log("No data available");
+                  return null;
+              }
+          }).catch((error) => {
+              console.error(error);
+          });
+          setLoading(true);
+        }, 1000);     
       });
   
       // Return the function to unsubscribe from the event so it gets removed on unmount
